Updates
Got the title of the post it note stack to stay centered whenever any edits are made to the title. Also designed the 'Done editing' button that indicates when the user is done editing the title. Also designed the add button at the bottom middle of the screen that will allow the user to add a new post it note stack.