Update
Renamed everything with 'stack' to 'section' for clarification that the initial dashboard contains post it note sections where each section contains multiple post it notes. Also worked on the done editing button regarding user input for the title of a post it note section. Also worked on preventing a separate post it note section title from becoming editable if the user has not clicked on the done editing button for a different post it note section.