Feature: Add feature that allows user to enter post it note section
When the user clicks on a post it note section, the user enters that section instead of entering edit mode for the section title. The useState currentlySelectedSection was renamed to selectedItem and setCurrentlySelectedSection was renamed to
setSelectedItem after making changes because the purpose of selectedItem is to track which item is currently selected so that the user is unable to edit other post it notes, including the section title post it note, when the user is already in edit mode
for one of the post it notes. selectedItem now tracks which item is selected in a section.