Refactor: Change "section" to "item" for some identifiers related to button bar component and move oneSectionIsEntered and setOneSectionIsEntered initialization from SelectedProvider.js to Dashboard.jsx
The word "section" was changed to "item" to generalize the button bar component so that it could be reused for both the post it note section array and post it note array that will exist in each section. Moved oneSectionIsentered and
setOneSectionIsEntered from SelectedProvider.js to Dashboard.jsx because of encountering issues involving passing oneSectionIsEntered to the ItemsButtonBar component. Issues occured because oneSectionIsEntered could not be passed to ItemsButtonBar
component since setOneSectionIsEntered cannot be called outside of the provider wrapper. setOneSectionIsEntered also is moved to Dashboard.jsx since it is paired with oneSectionIsEntered for useState.