Fix: Fix red highlight on front post it note while in a section
Fixed a bug where the front post it note would highlight red when the cursor hovers above it while the user is in a section. Only the individual post it note should highlight red in delete mode while in a section. Also changed the format of which div
creates a red highlight around the post it note in IndividualPostItNote.jsx so that it is consistent with the fix in PostItNoteSection.jsx where the bug was discovered.