Feature: Change interaction with post it note title
Interaction with the post it note title no longer includes the use of a done button. When the user is done editing the title, the user clicks off of the textarea. The code that deals with the appearance of tasks when the textarea is focused on or not is also stored in a custom hook called useDetectOnFocus.js. This code is part of a custom hook in case this particular custom hook is reused, especially in consideration of the post it note title. As a result of removing the done button, every instance of the identifiers textareaIsSelected and setTextareaIsSelected is removed.