Refactor: Implementing BEM css style
Created SelectToDelete.css to contain the different css styling for different uses so that the SelectToDelete component can change style when it is used for different scenarios. As a result of implementing BEM, using the functions and variables from the custom hook useDetectMouseHoverOver are no longer used. Instead, to change the css styling of components when the cursor hovers above them, the :hover in the css files are used. This is done to keep the class names simple since implementing BEM with the useDetectMouseHoverOver became more difficult to read.