Feature: Remove feature of the appearance of a border when the user hovers above a task in edit mode and remove implementation of custom hooks
Removed the implementation of useDetectOnFocus and useDetectMouseHoverOver from ToDoTask. Instead, the change in appearance now uses :focus in ToDoTask.css. This is done to keep consistency with the changes in other components where useDetectMouseHover is no longer implemented. The appearance of the border when the user hovers above a task in edit mode is also removed because this feature did not seeme to look visually appealing.