Fix: Fix bug to delete a task
The selectable area to delete a task did not resize when the textbox containing the task expanded in size. This occured because the size of the selectable area did not have a reference, which was needed to access the properties of the selectable area. The reference was created and properly passed to the function handling its size.