Refactor: Modify file structure regarding the ItemsButtonBar directory
Moved AddItemButton directory outside of the ItemsButtonBar directory because the AddItemButton component is re-usable outside of the items button bar. Deleted the Buttons directory in the ItemsButtonBar directory because the Buttons directory was created to group multiple directories together, but now that only the DeleteItemButton directory is within the ItemsButtonBar directory, the Buttons directory is no longer needed.