Refactor: Change AddItemButton component appearance regarding reusability
Styled the add item button to appear in a specific way for different uses. One use is when the add item button adds a new post it note, whether it is a post it note section or post it note to do list, and the other use is when the user adds a new task.