Fix: Fix bug where the done editing button stays on screen when the program is in delete mode
The done editing button stayed appearing on screen when the user entered delete mode to delete post it notes, meaning the user had access to leaving edit mode while the in delete mode. Pressing the done button while still in delete mode is not supposed to be allowed as it does not account for changes in state for different useState variables. This bug is fixed as the done editing button no longer appears while in delete mode.