Fix: Fix outdated number of tasks bug
When a section had one or more tasks, after all the tasks were removed, if the user returned to the dashboard then re-entered the same section, the section would have a single task. The method in updating the local storage was switched from checking if the array has no items to creating a useState variable that indicates if the useState variable that contains the array is updated and this fixed the bug.