Fix: Fix bug regarding difficulty of scrolling in post it note to do list
Previously had issue scrolling vertically in a post it note to do list if the cursor was hovering above a task, unless the scroll bar is first clicked. Issue occured while SelectToDelete appeared as a componented in front of the task. Updated SelectToDelete so that the content of the task appeared as a part of the SelectToDelete component instead of as a component that appeared in front of the other component containing the task. This is while SelectToDelete remained the same for deleting post it notes and this was achieved by creating a conditional to determine how the SelectToDelete would appear. Updated ToDoTask to properly resize the textarea when in edit for the task.