Refactor: Delete unecessary code
Code focusing on resizing a component relative to the size of a textarea is no longer needed. This feature was specifically for the tasks when the SelectToDelete component appeared in front of the tasks. However, the SelectToDelete and the ToDoTask components were modified to where the SelectToDelete component no longer appears in front of the textarea containing the task.