Fix: Fix bug where the content showing the task overlaps a little with the bullet point
After fixing the previous issue with SelectToDelete for post it notes, the content for the task overlapped over the bullet point since the change in css code in SelectToDelete.css affected the way the task appeared. Updated SelectToDelete.jsx to contain dynamic class names so that the css styling would change based on the type of item SelectToDelete dealt with.