Feature: Update local storage interaction with deleting items
For the content that is removed from removing a post it note section or post it note to do list, the content inside each is also removed from the local storage. The id representing the section itself or post it note to do list itself is also removed from the local storage.