Fix: Fix task display on display mode
Previously, if a task is created and has not been edited, after exiting edit mode, the height of the div that represents the task display is shorter than it should be. Updated SelectToDelete.jsx to adjust the initial size of the task when displayed if it has not been edited.