Fix and Refactor: Fix bug causing tasks and the to do list header to overlap when first loading the program and entering a section, and create custom hook for adjusting textarea
Previously encountered error where the tasks and to do list header would overlap. Updated ToDoListHeader and ToDoTask to give the div, that contained the content, height so that the program recognized that the height of the div, when the program is first loaded, represents the content height. This prevented the multiple divs from overlapping since before, the program did not recognize the divs' heights. Also created a custom hook to encapsulate the repeated logic of adjusting any textareas at specific instances.