Fix: Fix bug regarding entering a section
Previously, when entering a section closer to the top of the screen while scrolled down any amount, part of or all of a black screen would appear since the window did not reposition to the top of the webpage. Updated SectionUtils.js to account for this by repositioning to the top of the webpage right after clicking a section to enter.