Fix: Fix textarea resizing error
Previously, when inputting into a textarea for the first time, whether the textarea is related to a task, header, or title, the textarea slightly shifts in height. This occured because the initial textarea height is not equivalent to the textarea height after it is typed into. Fixed this resizing error by updating the custom hook useAdjustTextareaHeight and the files that rely on that custom hook.