Style and Fix: Position underline lower regarding the header and slightly readjusted the textarea height of the header
Placed the underline a little lower so that certain letters like "g" and "j" do not touch the underline after exiting edit mode. Readjusted the textarea height of the header when the header has no content so that the textarea does not have a noticeable change in height when the user starts typing in the textarea.