Fix: Fix bug regarding local storage with saved text
Previously, local storage did not properly save the text as after the program refreshes or a textarea re-renders, the text returns to an empty string. Updated the first useEffect in useLocalStoredTextarea to save the text in a useState variable if the locally stored text from the local storage is not an empty string and is not equal to null. This prevents the text to get set to the outdated text value, which is initially always an empty string as the useEffect with an empty array dependency runs twice, the second time being when the variable locallyStoredText is an empty string due to useEffect with the useState variable text as a dependency setting the local storage at the local key to an empty string since it uses the initial value of text, which is always an empty string. Also removed unecessary code for the useEffect that has the useState variable text as the dependency.