Fix: Fix bugs regarding ability to delete a section
Previously, if cursor hovered over the pin image, a section could not be deleted. Removed the z-index for the pin image to fix this issue. Also placed the pin image at a lower position so that the red highlight that surrounds a post it note section when the user is hovering over a section to delete does not noticeably overlap the pin.